+++ input.c
@@ -486,11 +486,13 @@ execute_binding(struct seat *seat, struct terminal *term,
         return true;
 
     case BIND_ACTION_THEME_SWITCH_1:
-        term_theme_switch_to_1(term);
+    case BIND_ACTION_THEME_SWITCH_DARK:
+        term_theme_switch_to_dark(term);
         return true;
 
     case BIND_ACTION_THEME_SWITCH_2:
-        term_theme_switch_to_2(term);
+    case BIND_ACTION_THEME_SWITCH_LIGHT:
+        term_theme_switch_to_light(term);
         return true;
 
     case BIND_ACTION_THEME_TOGGLE: