+++ terminal.h
@@ -994,8 +994,8 @@ void term_enable_size_notifications(struct terminal *term);
 void term_disable_size_notifications(struct terminal *term);
 void term_send_size_notification(struct terminal *term);
 
-void term_theme_switch_to_1(struct terminal *term);
-void term_theme_switch_to_2(struct terminal *term);
+void term_theme_switch_to_dark(struct terminal *term);
+void term_theme_switch_to_light(struct terminal *term);
 void term_theme_toggle(struct terminal *term);
 
 static inline void term_reset_grapheme_state(struct terminal *term)