+++ CHANGELOG.md
@@ -74,6 +74,8 @@
   `foot` and `footclient`), allowing you to set a custom toplevel
   tag. The compositor must implement the new `xdg-toplevel-tag-v1`
   Wayland protocol ([#2212][2212]).
+* `[colors-dark]` section to `foot.ini`. Replaces `[colors]`.
+* `[colors-light]` section to `foot.ini`. Replaces `[colors2]`.
 
 [2212]: https://codeberg.org/dnkl/foot/issues/2212
 
@@ -93,6 +95,11 @@
 
 
 ### Deprecated
+
+* `[colors]` section in `foot.ini`. Use `[colors-dark]` instead.
+* `[colors2]` section in `foot.ini`. Use `[colors-light]` instead.
+
+
 ### Removed
 ### Fixed